+++ examples/bss_example.py
@@ -26,8 +26,15 @@ The method implemented is described in the following publication
     J. Jansky, Z. Koldovsky, and N. Ono *A computationally cheaper method for blind speech
     separation based on AuxIVA and incomplete demixing transform*, Proc. IEEE, IWAENC, 2016.
 
+4) Fast Multichannel Nonnegative Matrix Factorization (FastMNMF)
+The method implemented is described in the following publication
+
+    K. Sekiguchi, A. A. Nugraha, Y. Bando, K. Yoshii, *Fast Multichannel Source 
+    Separation Based on Jointly Diagonalizable Spatial Covariance Matrices*, EUSIPCO, 2019.
+
 All the algorithms work in the STFT domain. The test files were extracted from the
 `CMU ARCTIC <http://www.festvox.org/cmu_arctic/>`_ corpus.
+
 
 Depending on the input arguments running this script will do these actions:.
 
@@ -58,7 +65,7 @@ wav_files = [
 
 if __name__ == '__main__':
 
-    choices = ['ilrma', 'auxiva', 'sparseauxiva']
+    choices = ['ilrma', 'auxiva', 'sparseauxiva', 'fastmnmf']
 
     import argparse
     parser = argparse.ArgumentParser(description='Demonstration of blind source separation using '
@@ -82,6 +89,9 @@ if __name__ == '__main__':
 
     ## Prepare one-shot STFT
     L = args.block
+    hop = L // 2
+    win_a = pra.hann(L)
+    win_s = pra.transform.compute_synthesis_window(win_a, hop)
 
     ## Create a room with sources and mics
     # Room dimensions in meters
@@ -139,13 +149,15 @@ if __name__ == '__main__':
         global SDR, SIR
         from mir_eval.separation import bss_eval_sources
         ref = np.moveaxis(separate_recordings, 1, 2)
-        y = pra.transform.synthesis(Y, L, L, zp_back=L//2, zp_front=L//2).T
-        sdr, sir, sar, perm = bss_eval_sources(ref[:,:y.shape[1]-L//2,0], y[:,L//2:ref.shape[1]+L//2])
+        y = pra.transform.synthesis(Y, L, hop, win=win_s)
+        y = y[L-hop: , :].T
+        m = np.minimum(y.shape[1], ref.shape[1])
+        sdr, sir, sar, perm = bss_eval_sources(ref[:, :m, 0], y[:, :m])
         SDR.append(sdr)
         SIR.append(sir)
 
     ## STFT ANALYSIS
-    X = pra.transform.analysis(mics_signals.T, L, L, zp_front=L//2, zp_back=L//2)
+    X = pra.transform.analysis(mics_signals.T, L, hop, win=win_a)
 
     t_begin = time.perf_counter()
 
@@ -158,6 +170,10 @@ if __name__ == '__main__':
     elif bss_type == 'ilrma':
         # Run ILRMA
         Y = pra.bss.ilrma(X, n_iter=30, n_components=2, proj_back=True,
+                          callback=convergence_callback)
+    elif bss_type == 'fastmnmf':
+        # Run FastMNMF
+        Y = pra.bss.fastmnmf(X, n_iter=100, n_components=8, n_src=2,
                           callback=convergence_callback)
     elif bss_type == 'sparseauxiva':
         # Estimate set of active frequency bins
@@ -173,10 +189,12 @@ if __name__ == '__main__':
     print("Time for BSS: {:.2f} s".format(t_end - t_begin))
     
     ## STFT Synthesis
-    y = pra.transform.synthesis(Y, L, L, zp_front=L//2, zp_back=L//2).T
+    y = pra.transform.synthesis(Y, L, hop, win=win_s)
 
     ## Compare SDR and SIR
-    sdr, sir, sar, perm = bss_eval_sources(ref[:,:y.shape[1]-L//2,0], y[:,L//2:ref.shape[1]+L//2])
+    y = y[L-hop:, :].T
+    m = np.minimum(y.shape[1], ref.shape[1])
+    sdr, sir, sar, perm = bss_eval_sources(ref[:, :m, 0], y[:, :m])
     print('SDR:', sdr)
     print('SIR:', sir)