+++ pyroomacoustics/__init__.py
@@ -101,6 +101,9 @@ Available subpackages
 :py:obj:`pyroomacoustics.doa`
     Direction of arrival finding algorithms
 
+:py:obj:`pyroomacoustics.phase`
+    Phase-related processing
+
 :py:obj:`pyroomacoustics.transform`
     Block frequency domain processing tools
 
@@ -137,6 +140,7 @@ from . import experimental
 from . import datasets
 from . import bss
 from . import denoise
+from . import phase
 
 import warnings
 warnings.warn(