+++ pyroomacoustics/bss/tests/test_bss.py
@@ -97,6 +97,9 @@ def freq_bss(algo='auxiva', L=256, **kwargs):
     elif algo == 'overiva':
         Y = pra.bss.auxiva(X, n_src=1, n_iter=30, proj_back=True, **kwargs)
         max_mse = 0.5
+    elif algo == 'fastmnmf':
+        Y = pra.bss.fastmnmf(X, n_src=2, n_iter=30, n_components=16)
+        max_mse = 1e-1
 
     ## STFT Synthesis
     if algo == 'overiva':
@@ -149,6 +152,11 @@ class TestBSS(unittest.TestCase):
     def test_bss_overiva(self):
         for block in L:
             freq_bss(algo='overiva', L=block)
+
+    # Test fastmnmf with frame lengths [256, 512, 1024, 2048, 4096]
+    def test_bss_fastmnmf(self):
+        for block in L:
+            freq_bss(algo='fastmnmf', L=block)
 
 
 if __name__ == '__main__':