+++ pyroomacoustics/experimental/rt60.py
@@ -1,7 +1,40 @@
+# Copyright (c) 2018-2019 Robin Scheibler
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+"""
+RT60 Measurement Routine
+========================
+
+Automatically determines the reverberation time of an impulse response
+using the Schroeder method [1]_.
+
+References
+----------
+
+.. [1] M. R. Schroeder, "New Method of Measuring Reverberation Time,"
+    J. Acoust. Soc. Am., vol. 37, no. 3, pp. 409-412, Mar. 1968.
+"""
 import numpy as np
 
+
 def measure_rt60(h, fs=1, decay_db=60, plot=False, rt60_tgt=None):
-    '''
+    """
     Analyze the RT60 of an impulse response. Optionaly plots some useful information.
 
     Parameters
@@ -20,7 +53,7 @@ def measure_rt60(h, fs=1, decay_db=60, plot=False, rt60_tgt=None):
     rt60_tgt: float
         This parameter can be used to indicate a target RT60 to which we want
         to compare the estimated value.
-    '''
+    """
 
     h = np.array(h)
     fs = float(fs)
@@ -42,7 +75,6 @@ def measure_rt60(h, fs=1, decay_db=60, plot=False, rt60_tgt=None):
 
     # after decay
     i_decay = np.min(np.where(-5 - decay_db - energy_db > 0)[0])
-    e_decay = energy_db[i_decay]
     t_decay = i_decay / fs
 
     # compute the decay time
@@ -60,21 +92,24 @@ def measure_rt60(h, fs=1, decay_db=60, plot=False, rt60_tgt=None):
         power_db -= np.max(power_db)
 
         # time vector
-        get_time = lambda x, fs: np.arange(x.shape[0]) / fs - i_5db / fs
+        def get_time(x, fs):
+            return np.arange(x.shape[0]) / fs - i_5db / fs
+
         T = get_time(power_db, fs)
 
         # plot power and energy
-        plt.plot(T, power_db, label='Power')
-        plt.plot(get_time(energy_db, fs), energy_db, label='Energy')
+        plt.plot(T, power_db, label="Power")
+        plt.plot(get_time(energy_db, fs), energy_db, label="Energy")
 
         # now the linear fit
-        decay_rate = (5 + decay_db) / decay_time
-        plt.plot([0, est_rt60], [e_5db, -65], '--', label='Linear Fit')
-        plt.plot(T, np.ones_like(T) * -60, '--', label='-60 dB')
-        plt.vlines(est_rt60, energy_db_min, 0, linestyles='dashed', label='Estimated RT60')
+        plt.plot([0, est_rt60], [e_5db, -65], "--", label="Linear Fit")
+        plt.plot(T, np.ones_like(T) * -60, "--", label="-60 dB")
+        plt.vlines(
+            est_rt60, energy_db_min, 0, linestyles="dashed", label="Estimated RT60"
+        )
 
         if rt60_tgt is not None:
-            plt.vlines(rt60_tgt, energy_db_min, 0, label='Target RT60')
+            plt.vlines(rt60_tgt, energy_db_min, 0, label="Target RT60")
 
         plt.legend()